requirements.txt
discord.py==2.3.2
google-api-python-client==2.97.0
python-dotenv==1.0.0
aiohttp==3.9.3
PyNaCl==1.5.0
oauth2client<4.0.0
